requirements.txt
# Movie Recommendation System - Dependencies
# Updated for FAISS-based similarity search

# Core data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0  # Parquet support

# Machine learning & NLP
scikit-learn>=1.3.0
sentence-transformers>=2.2.0  # Neural embeddings (SBERT)
faiss-cpu>=1.7.4  # Similarity search (use faiss-gpu for GPU)

# Data validation
pandera>=0.17.0

# Web backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Frontend
streamlit>=1.28.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0

# Development & testing
pytest>=7.4.0
httpx>=0.25.0  # For FastAPI testing
